+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 # Change Log
 
-## 1.1.24 -- Developing...
+## 1.1.25 -- Developing...
+
+## 1.1.24
 * Updating the datagrid when the props are updated
 * Slider component added.